Extract `concatBytes` utility
